+++ src/main/java/sttDB/service/interproServices/TrapidInterproParser.java
@@ -1,22 +1,40 @@
 package sttDB.service.interproServices;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import sttDB.exception.InterproParsingException;
 
 import java.io.*;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class TrapidInterproParser implements InterproParser {
 
     private BufferedReader reader;
 
+    List<LineItems> items;
+
     @Override
     public List<LineItems> parse() throws InterproParsingException {
         if (reader == null)
             throw new InterproParsingException("File to parse not set, call 'setFileToParse first'");
-        return null;
+        items = reader.lines().skip(1).map(this::parseLine).collect(Collectors.toList());
+        return items;
+    }
+
+    private LineItems parseLine(String line) {
+        List<String> parts = Arrays.asList(line.split("\\s+"));
+        String trinityID = parts.get(1);
+        String interproID = parts.get(2);
+        String description = getDescription(parts);
+        return new LineItems(trinityID, interproID, description);
+    }
+
+    private String getDescription(List<String> parts) {
+        return parts.stream()
+                .skip(3)
+                .collect(Collectors.joining(" "));
     }
 
     @Override
